refactor:feature_search -> 참여,생성시 보증금차감 및 계좌엔티티 업데이트

+++ src/main/java/com/zerobase/challengeproject/challenge/domain/dto/GetChallengeDto.java
@@ -18,8 +18,8 @@ public class GetChallengeDto {
     private CategoryType categoryType;
     private Integer participant;
     private String description;
-    private Integer min_deposit;
-    private Integer max_deposit;
+    private Long min_deposit;
+    private Long max_deposit;
     private String standard;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
