+++ src/main/java/com/zerobase/challengeproject/challenge/domain/dto/ParticipationChallengeDto.java
@@ -17,8 +17,8 @@ public class ParticipationChallengeDto {
     private CategoryType categoryType;
     private Integer participant;
     private String description;
-    private Integer min_deposit;
-    private Integer max_deposit;
+    private Long min_deposit;
+    private Long max_deposit;
     private String standard;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
